# Improve performance for det(F) where F is a LUfactor