I noticed that some of the vectorized operations in the determinants on LUFactors took a very long time. This devectorizes them for a significant speed up, especially for small matrices (600x speedup). These are used in the calculation of determinants for normal matrices so it is good to get them fast.

Benchmark code
```julia
function f(N, n)
    A = rand(n,n)
    F = lufact(A)
    @time for i = 1:N
        logabsdet(F)
    end
     @time for i = 1:N
        det(F)
    end
end

for elty in (Float64, Complex128)
    println("eltype $elty:")
    println("n = 4")
    f(10^5, 4)
    println("n = 100")
    f(10^5, 100)
end
```
### Master
```
eltype Float64:
n = 4
  0.604570 seconds (2.50 M allocations: 505.066 MB, 5.28% gc time)
  0.579472 seconds (2.20 M allocations: 477.600 MB, 4.88% gc time)
n = 100
  0.886912 seconds (2.50 M allocations: 810.242 MB, 3.91% gc time)
  0.651776 seconds (2.20 M allocations: 553.894 MB, 4.27% gc time)
eltype Complex{Float64}:
n = 4
  0.603211 seconds (2.50 M allocations: 505.066 MB, 5.09% gc time)
  0.577971 seconds (2.20 M allocations: 477.600 MB, 4.83% gc time)
n = 100
  0.849031 seconds (2.50 M allocations: 810.242 MB, 3.81% gc time)
  0.650605 seconds (2.20 M allocations: 553.894 MB, 4.08% gc time)
```

### PR
```
eltype Float64:
n = 4
  0.007293 seconds
  0.000908 seconds
n = 100
  0.173780 seconds
  0.011353 seconds
eltype Complex{Float64}:
n = 4
  0.007362 seconds
  0.000880 seconds
n = 100
  0.165616 seconds
  0.011188 seconds
```