@KristofferC 2015-08-05
I think AppVeyor needs a restart.  Not sure what the errors on Travis are. Anyone?