@tkelman 2015-08-05
Bugs for everyone. Travis was one `fatal: unable to connect to github.com:` and one `fatal: unable to access 'https://github.com/JuliaLang/JuliaDoc.git/': gnutls_handshake() failed: A TLS packet with unexpected length was received.`, and appveyor was https://github.com/appveyor/ci/issues/353. Can you rebase and force push here?