@KristofferC 2015-08-06
Rebased and green now. From my point of view, this is good to go.